Updated Radar so that correct routing button is selected and resolution is changed for PPM vs outer modes

+++ plugins/radar/radarmeter.h
@@ -135,7 +135,7 @@ class RadarMeter : public wxWindow
 
         bool m_bShowLevels;
 
-        float m_dMindB;
+        double m_dMindB;
         float m_dResolution;
 
         double m_dAngle;
